Move subgraph py to main module
