+++ pdr_backend/util/subgraph.py
@@ -57,7 +57,7 @@ query {
 
 import os
 import requests
-from typing import Optional, Dict
+from typing import Optional, Dict, List
 
 from enforce_typing import enforce_types
 from web3 import Web3
@@ -129,6 +129,37 @@ def query_subgraph(subgraph_url: str, query: str) -> Dict[str, dict]:
     result = request.json()
     return result
 
+
+def query_pending_payouts(addr: str) -> List[int]:
+    chunk_size = 1000
+    offset = 0
+    timestamps: List[int] = []
+
+    while True:
+        query = """
+        {
+                predictPredictions(
+                    where: {user: "%s", payout: null}
+                ) {
+                    id
+                    timestamp
+                }
+        }
+        """ % (addr)
+
+        offset += chunk_size
+        try:
+            result = query_subgraph(subgraph_url, query)
+            if not "data" in result:
+                print("No data in result")
+                break
+            predict_predictions = result["data"]["predictPredictions"]
+            timestamps_query = [i["timestamp"] for i in predict_predictions]
+            timestamps.extend(timestamps_query)
+        except Exception as e:
+            print("An error occured", e)
+
+    return timestamps
 
 @enforce_types
 def query_predictContractss(
